LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-video-element/video-loading-lazy-poster-when-visible-expected.txt



FAIL Video with loading=lazy does not fetch poster while not visible, then fetches poster once scrolled into view assert_equals: Poster should not be fetched when not visible expected 0 but got 1
